Fix OCR PDF bugs, modify encoding, change savedir
The OCR pdf bug is found: for some pdf pages, the script will fail. For
such pages, we convert it to png instead. Also, we use binary search
suggested by jzc to minimize the number of pages to convert to png while
maintaining the efficicency.

The encoding mechanism of directory/files are modified to use the
`Jzchash` function. Also, we avoid using relative path by converting
path to full path each time calling in the function.

The saving directory is migrated to /tmp, so the intermidiate files
won't be left up on the user's device.